Add custom type face for cards
The views retrieve the font family specified by app, if none is found, default configuration is used.

The font styling is configured as follor:

a Generic font is specified, a Title and a Message.

For instance, if generic font is specified but no title nor message, all the texts in the card will set the generic font.

Other case is if the title is specified but no generic nor message, only a custom font for title text will be set.